Refractor pgdtools
- remove proxy list
- move inner classes into individual files in their own folders
  - keep pgdtools.py cleaner
  - allow for keeping all files together cleaner
  - use imports in `__init__.py` in `sub_tools`
- Restructure tests to follow along with everything.

There are still a bunch of garbage routines in the files, these will need to be cleaned up next and stuff we don't want deleted.
