+++ src/pgdtools/pgdtools.py
@@ -1,12 +1,16 @@
-"""Tools to read and work with the presolar grain database."""
+"""Main PresolarGrains class that handles the overall database stuff.
 
+All sub functions and tools live in the `sub_tools` folder and are imported here."""
+
+from enum import Enum
 from typing import List, Tuple, Union
 
 from iniabu import ini
 import numpy as np
 import pandas as pd
 
-from pgdtools import db, utilities as utils
+from pgdtools import db
+from pgdtools.sub_tools import Grain
 
 
 class PresolarGrains:
@@ -26,145 +30,52 @@ class PresolarGrains:
         Load the default database into self.db and self._db as a backup.
         """
         try:
-            filepath = db.current()["sic"]
+            curr_db = db.current()
         except FileNotFoundError:
             print("No default database found, downloading latest versions...")
             db.update()
-            filepath = db.current()["sic"]
+            curr_db = db.current()
 
-        self.db = pd.read_csv(filepath, index_col=0)
+        keys = curr_db.keys()
+        if not keys:
+            raise ValueError("No database found. Try to update the database.")
+
+        filepaths = []
+        for key in keys:
+            filepaths.append(curr_db[key])
+
+        dfs = [pd.read_csv(filepath, index_col=0) for filepath in filepaths]
+        self.db = pd.concat(dfs)
         self._db = self.db.copy(deep=True)
 
-    class Grain:
-        """Class to represent a single grain."""
+    class DataBase(Enum):
+        """Enum to represent the different databases.
 
-        def __init__(self, parent: "PresolarGrains", id: str):
-            """Initialize the class for a single grain.
+        The name of each enum is something that should make sense to the user,
+        the value the 3 letter abbreviation that each PGD Grain Name starts with.
+        """
 
-            :param parent: Parent class, must be of type ``PresolarGrains``.
-            :param id: PGD ID of the grain to initialize, or list of IDs.
+        SiC = "SiC"
+        Gra = "Gra"
 
-            :raises TypeError: Parent class is not of type ``PresolarGrains``.
-            """
-            if not isinstance(parent, PresolarGrains):
-                raise TypeError("Parent class must be of type PresolarGrains.")
+    # SUB TOOL ACCESS #
 
-            self.parent = parent
-            self._id = id
+    def grain(self, grain_id: Union[str, List[str]]):
+        """Return a single grain object for a specific ID.
 
-            self._entry = self.parent.db.loc[self._id]
+        :param grain_id: PGD ID of the grain or a list of IDs
 
-        @property
-        def entry(self) -> Union[pd.Series, pd.DataFrame]:
-            """Return the whole data set of this entry.
+        :raises ValueError: ID not found in database.
+        """
+        if isinstance(grain_id, str):
+            grain_id = [grain_id]
 
-            If only one entry is in dataframe, a pandas series is returned.
-
-            :return: Data set of the grain.
-            """
-            if self._entry.shape[0] == 1:  # A Series
-                ret_val = self._entry.iloc[0]
-            else:
-                ret_val = self._entry
-            return ret_val
-
-        @property
-        def id(self):
-            """Return the PGD ID of the grain."""
-            return utils.return_list_simplifier(self._id)
-
-        @property
-        def pgd_type(self) -> Union[Tuple[str, str], Tuple[pd.Series, pd.Series]]:
-            """Return the PGD type and subtype of the grain(s).
-
-            :return: PGD type and subtype. If no subtype, returns ``NaN``.
-            """
-            subtype = self._entry["PGD Subtype"]
-            if isinstance(subtype, pd.Series):
-                subtype.replace({np.nan: None}, inplace=True)
-            elif subtype == np.nan:
-                subtype = None
-            return (
-                utils.return_list_simplifier(self._entry["PGD Type"]),
-                utils.return_list_simplifier(subtype),
-            )
-
-        @property
-        def probabilities(self) -> dict:
-            """Return the probabilities for all types for a given grain.
-
-            The probabilities are returned as a dictionary with key values of all types,
-            namely M, X, Y, Z, AB, C, D, and N. The values are the probabilities for
-            the given grain to be of the given type. See the paper for detail.
-
-            :return: Probabilities for all types.
-            """
-            ret_dict = {}
-            for pgd_type in ["M", "X", "Y", "Z", "AB", "C", "D", "N"]:
-                ret_dict[pgd_type] = utils.return_list_simplifier(
-                    self._entry[f"p({pgd_type})"]
-                )
-            return ret_dict
-
-        @property
-        def reference(self) -> Union[str, pd.Series]:
-            """Return the reference of the grain."""
-            return utils.return_list_simplifier(self._entry["Reference"])
-
-        @property
-        def source(self) -> str:
-            """Return the source of the grain."""
-            return str(utils.return_list_simplifier(self._entry["Source"]))
-
-        def correlation(self, iso1: str, iso2: str) -> Union[float, pd.Series]:
-            """Return the correlation between two isotopes.
-
-            If no correlation is recorded, zero is returned (no correlation).
-
-            :param iso1: First isotope.
-            :param iso2: Second isotope.
-
-            :return: Correlation between the two isotopes.
-            """
-            corr = self._entry[self.parent.header_correlation(iso1, iso2)]
-            if isinstance(corr, pd.Series):
-                corr.replace({np.nan: 0}, inplace=True)
-            elif corr == np.nan:
-                corr = 0
-            return corr
-
-        def value(
-            self, iso1: str, iso2: str
-        ) -> Union[
-            Tuple[float, Union[float, Tuple[float, float]], bool],
-            Tuple[pd.Series, Union[pd.Series, Tuple[pd.Series, pd.Series]], bool],
-        ]:
-            """Return the value stored isotope ratio.
-
-            If no value is recorded, ``NaN`` is returned.
-
-            :param iso1: Nominator isotope.
-            :param iso2: Denominator isotope.
-
-            :return: value, sigma or (sigma+, sigma-), is_delta?
-            """
-            hdr, err, is_delta = self.parent.header_ratio(iso1, iso2)
-            if isinstance(err, Tuple):
-                errors = (
-                    utils.return_list_simplifier(self._entry[err[0]]),
-                    utils.return_list_simplifier(self._entry[err[1]]),
-                )
-            else:
-                errors = utils.return_list_simplifier(self._entry[err])
-            return utils.return_list_simplifier(self._entry[hdr]), errors, is_delta
+        grain_ids_not_found = [gid for gid in grain_id if gid not in self.db.index]
+        if grain_ids_not_found:
+            raise ValueError(f"Grain IDs {grain_ids_not_found} not found in database.")
+        return Grain(self, grain_id)
 
     # PROPERTIES #
-
-    @property
-    def grain(self):
-        """Return instance for (a) specific grain ID(s)."""
-        valid_keys = list(self.db.index)
-        return utils.ProxyList(self, self.Grain, valid_keys)
 
     @property
     def reference(self) -> pd.DataFrame:
