+++ src/pgdtools/sub_tools/__init__.py
@@ -1,0 +1,5 @@
+"""Sub-tools for all the Presolar Grain tools."""
+
+from .grain import Grain
+
+__all__ = ["Grain"]
