+++ tests/conftest.py
@@ -26,15 +26,20 @@ def pgd(pgd_setup):
 def pgd_setup(tmpdir_home, data_files_dir):
     """Copy and create the required files for reading the PGD."""
     sic_db_name = "PGD_SiC_2023-10-30.csv"
+    gra_db_name = "PGD_Gra_2024-05-13.csv"
 
     db_json = tmpdir_home.joinpath("config/db.json")
     db_json.write_text(data_files_dir.joinpath("db.json").read_text())
 
     sic_db = tmpdir_home.joinpath(f"csv/{sic_db_name}")
     sic_db.write_bytes(data_files_dir.joinpath(sic_db_name).read_bytes())
+    gra_db = tmpdir_home.joinpath(f"csv/{gra_db_name}")
+    gra_db.write_bytes(data_files_dir.joinpath(gra_db_name).read_bytes())
 
     current_file = tmpdir_home.joinpath("current.json")
-    current_file.write_text(f'{{"sic": "{str(sic_db.absolute())}"}}')  # noqa: B907
+    current_file.write_text(
+        f'{{"sic": "{str(sic_db.absolute())}", "gra": "{str(gra_db.absolute())}"}}'
+    )  # noqa: B907
 
 
 @pytest.fixture(autouse=True)
